+++ SampleCode/StdDriver/OTG_Dual_Role_UMAS/main.c
@@ -238,7 +238,7 @@ void put_dump (
     int i;
 
 
-    printf("%08x ", addr);
+    printf("%08x ", (UINT)addr);
 
     for (i = 0; i < cnt; i++)
         printf(" %02x", buff[i]);
@@ -476,7 +476,7 @@ void USBH_Process()
                 break;
             }
             sect = p2 + 1;
-            printf("Sector:%d\n", p2);
+            printf("Sector:%d\n", (INT)p2);
             for (buf=(unsigned char*)Buff1, ofs = 0; ofs < 0x200; buf+=16, ofs+=16)
                 put_dump(buf, ofs, 16);
             break;
@@ -553,9 +553,9 @@ void USBH_Process()
                 printf("FAT type = FAT%d\nBytes/Cluster = %d\nNumber of FATs = %d\n"
                        "Root DIR entries = %d\nSectors/FAT = %d\nNumber of clusters = %d\n"
                        "FAT start (lba) = %d\nDIR start (lba,cluster) = %d\nData start (lba) = %d\n\n...",
-                       ft[fs->fs_type & 3], fs->csize * 512UL, fs->n_fats,
-                       fs->n_rootdir, fs->fsize, fs->n_fatent - 2,
-                       fs->fatbase, fs->dirbase, fs->database
+								       ft[fs->fs_type & 3], (UINT)(fs->csize * 512UL), fs->n_fats,
+                       fs->n_rootdir, (INT)fs->fsize, (INT)(fs->n_fatent - 2),
+                       (INT)fs->fatbase, (INT)fs->dirbase, (INT)fs->database
                       );
                 acc_size = acc_files = acc_dirs = 0;
 #if _USE_LFN
@@ -570,9 +570,9 @@ void USBH_Process()
                 }
                 printf("\r%d files, %d bytes.\n%d folders.\n"
                        "%d KB total disk space.\n%d KB available.\n",
-                       acc_files, acc_size, acc_dirs,
-                       (fs->n_fatent - 2) * (fs->csize / 2), p2 * (fs->csize / 2)
-                      );
+                       acc_files, (INT)acc_size, acc_dirs,
+                       (INT)((fs->n_fatent - 2) * (fs->csize / 2)), (INT)(p2 * (fs->csize / 2))
+                );
                 break;
             case 'l' :  /* fl [<path>] - Directory listing */
                 while (*ptr == ' ') ptr++;
@@ -603,7 +603,7 @@ void USBH_Process()
                            (Finfo.fattrib & AM_SYS) ? 'S' : '-',
                            (Finfo.fattrib & AM_ARC) ? 'A' : '-',
                            (Finfo.fdate >> 9) + 1980, (Finfo.fdate >> 5) & 15, Finfo.fdate & 31,
-                           (Finfo.ftime >> 11), (Finfo.ftime >> 5) & 63, Finfo.fsize, Finfo.fname);
+                           (Finfo.ftime >> 11), (Finfo.ftime >> 5) & 63, (INT)Finfo.fsize, Finfo.fname);
 #if _USE_LFN
                     for (p2 = strlen(Finfo.fname); p2 < 14; p2++)
                         printf(" ");
@@ -612,9 +612,9 @@ void USBH_Process()
                     printf("\n");
 #endif
                 }
-                printf("%4d File(s),%10d bytes total\n%4d Dir(s)", s1, p1, s2);
+                printf("%4d File(s),%10d bytes total\n%4d Dir(s)", s1, (INT)p1, s2);
                 if (f_getfree(ptr, (DWORD*)&p1, &fs) == FR_OK)
-                    printf(", %10d bytes free\n", p1 * fs->csize * 512);
+                    printf(", %10d bytes free\n", (INT)(p1 * fs->csize * 512));
                 break;
 
 
@@ -633,7 +633,7 @@ void USBH_Process()
                 res = f_lseek(&file1, p1);
                 put_rc(res);
                 if (res == FR_OK)
-                    printf("fptr=%d(0x%lX)\n", file1.fptr, file1.fptr);
+                    printf("fptr=%d(0x%lX)\n", (INT)file1.fptr, file1.fptr);
                 break;
 
             case 'd' :  /* fd <len> - read and dump file from current fp */
@@ -690,7 +690,7 @@ void USBH_Process()
                 }
                 p1 = get_ticks()-t0;
                 if (p1)
-                    printf("%d bytes read with %d kB/sec.\n", p2, ((p2 * 100) / p1)/1024);
+                    printf("%d bytes read with %d kB/sec.\n", (INT)p2, (INT)(((p2 * 100) / p1)/1024));
                 break;
 
             case 'w' :  /* fw <len> <val> - write file */
@@ -721,7 +721,7 @@ void USBH_Process()
                 }
                 p1 = get_ticks()-t0;
                 if (p1)
-                    printf("%d bytes written with %d kB/sec.\n", p2, ((p2 * 100) / p1)/1024);
+                    printf("%d bytes written with %d kB/sec.\n", (INT)p2, (INT)(((p2 * 100) / p1)/1024));
                 break;
 
             case 'n' :  /* fn <old_name> <new_name> - Change file/dir name */
@@ -794,7 +794,7 @@ void USBH_Process()
                     p1 += s2;
                     if (res || s2 < s1) break;   /* error or disk full */
                 }
-                printf("\n%d bytes copied.\n", p1);
+                printf("\n%d bytes copied.\n", (INT)p1);
                 f_close(&file1);
                 f_close(&file2);
                 break;
@@ -850,7 +850,7 @@ void USBH_Process()
                     }
 
                     if ((p1 % 0x10000) == 0)
-                        printf("\n%d KB compared.", p1/1024);
+                        printf("\n%d KB compared.", (INT)(p1/1024));
                     printf(".");
                 }
                 if (s1 == 0)
